% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleView.R
\name{abdPlotTabsGUI}
\alias{abdPlotTabsGUI}
\title{Abundances barplots inside Tk tabs.}
\usage{
abdPlotTabsGUI(RclusTool.env = initParameters())
}
\arguments{
\item{RclusTool.env}{environment in which all global parameters, raw data and results are stored.}
}
\value{
None
}
\description{
Display the abundances barplot of a clustering.
}
\details{
abdPlotTabsGUI calls abdPlotTabs to display the abundances barplot of the clusterings of the current RclusTool.env$data.sample, inside the GUI.
}
\keyword{internal}
