% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchTab.R
\name{buildBatchTab}
\alias{buildBatchTab}
\title{Batch process tab}
\usage{
buildBatchTab(RclusTool.env)
}
\arguments{
\item{RclusTool.env}{environment in which data and intermediate results are stored.}
}
\value{
None
}
\description{
Generate the batch process tab of the \code{\link{RclusToolGUI}}, in which the user can choose and configure the classification method to apply on several datasets.
}
\details{
function to create the 'batchTab' for data clustering in batch process
}
\keyword{internal}
