% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessTab.R
\name{buildPreprocessTab}
\alias{buildPreprocessTab}
\title{build Preprocess tab}
\usage{
buildPreprocessTab(RclusTool.env)
}
\arguments{
\item{RclusTool.env}{environment in which data and intermediate results are stored.}
}
\value{
None
}
\description{
Generate the data preprocessing tab of the \code{\link{RclusToolGUI}}, in which the user can select, transform, filter or create variables.
}
\details{
function to build the 'preprocessTab' for data preprocessing (variables selection, transformation, creation, ...)
}
\keyword{internal}
