% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semisupTab.R
\name{buildSemisupTab}
\alias{buildSemisupTab}
\title{Semi-Supervised tab}
\usage{
buildSemisupTab(RclusTool.env)
}
\arguments{
\item{RclusTool.env}{environment in which data and intermediate results are stored.}
}
\value{
None
}
\description{
Generate the semi-supervised classification tab of the \code{\link{RclusToolGUI}}, in which the user can choose and configure the semi-supervised method to apply.
}
\details{
function to build the 'semisupTab' for semi-supervised classification
}
\keyword{internal}
