% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supTab.R
\name{buildsupTab}
\alias{buildsupTab}
\title{Supervised tab}
\usage{
buildsupTab(RclusTool.env)
}
\arguments{
\item{RclusTool.env}{: environment in which data and intermediate results are stored.}
}
\value{
None
}
\description{
Generate the supervised classification tab of the \code{\link{RclusToolGUI}}, in which the user can choose and configure the supervised method to apply.
}
\details{
function to create the 'supTab' for supervised classification
}
\keyword{internal}
