% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupTab.R
\name{buildUnsupTab}
\alias{buildUnsupTab}
\title{Unsupervised tab}
\usage{
buildUnsupTab(RclusTool.env)
}
\arguments{
\item{RclusTool.env}{: environment in which data and intermediate results are stored.}
}
\value{
None
}
\description{
This function generates the unsupervised classification tab of the \code{\link{RclusToolGUI}}, in which the user can choose and configure the clustering method to apply.
}
\details{
function to create the 'unsupTab' for data clustering
}
\keyword{internal}
