% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleCompute.R
\name{computeItemsGUI}
\alias{computeItemsGUI}
\title{GUI to estimate the number of cells in colonies for each cluster}
\usage{
computeItemsGUI(method.select = "K-means",
  RclusTool.env = RclusTool.env)
}
\arguments{
\item{method.select}{character vector specifying the name of the clustering result to use ('K-means' by default).}

\item{RclusTool.env}{environment in which all global parameters, raw data and results are stored.}
}
\value{
csv file containing the counts.
}
\description{
Open a Graphical User Interface allowing to choose cluster name and model file for the estimation and the saving of the number of cells in colonies.
}
\details{
computeItemsGUI opens a Graphical User Interface allowing to choose cluster name and model file for the estimation and the saving of the number of cells in colonies
}
\examples{
\donttest{ 

computeItemsGUI(method.select="K-means")
 
}
}
\keyword{internal}
