% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleCompute.R
\name{convNamesPairsToIndexPairs}
\alias{convNamesPairsToIndexPairs}
\title{Conversion of a set of names pairs to matrix of index pairs (2 columns)}
\usage{
convNamesPairsToIndexPairs(pair.list, full.name.set)
}
\arguments{
\item{pair.list}{list of names pairs (vector of 2 names).}

\item{full.name.set}{a vector of unique names.}
}
\value{
the 2-column matrix of pairs of indexes.
}
\description{
Performs conversion of list of names pairs to a matrix of index pairs
}
\details{
convNamesPairsToIndexPairs performs conversion of list of names pairs to a matrix of index pairs.
}
\examples{
x <- c("a"=10, "b"=20, "c"=30 )

ML <- list()
ML[[1]] <- c("a","c")
ML[[2]] <- c("c","d")

mustLink<-convNamesPairsToIndexPairs(ML,names(x))

}
\seealso{
\code{\link{computeSemiSupervised}}, \code{\link{KwaySSSC}}
}
\keyword{internal}
