% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleCompute.R
\name{convNamesToIndex}
\alias{convNamesToIndex}
\title{Conversion of element names to indexes}
\usage{
convNamesToIndex(name.sample, full.name.set)
}
\arguments{
\item{name.sample}{a vector of element names.}

\item{full.name.set}{a vector of unique names.}
}
\value{
the vector of indexes of the element named.
}
\description{
Performs conversion of an element name to its index inside a set of named objects
}
\details{
convNamesToIndex performs conversion of an element name to its index inside a set of named objects
}
\examples{
x <- c("a"=10, "b"=20, "c"=30 )

convNamesToIndex(c("b","a","c"),names(x))

}
\seealso{
\code{\link{computeSemiSupervised}}, \code{\link{KwaySSSC}}
}
\keyword{internal}
