% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleView.R
\name{cor.mtest}
\alias{cor.mtest}
\title{Correlation test.}
\usage{
cor.mtest(mat)
}
\arguments{
\item{mat}{correlation matrix}
}
\value{
matrix with p-values.
}
\description{
Display the abundances barplot of a clustering.
}
\details{
cor.mtest Correlation test on correlation matrix.
}
\references{
\url{https://www.sthda.com/french/wiki/visualiser-une-matrice-de-correlation-par-un-correlogramme}
}
\keyword{internal}
