% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleCompute.R
\name{countItemsGUI}
\alias{countItemsGUI}
\title{GUI to manually count the number of cells in colonies}
\usage{
countItemsGUI(RclusTool.env = initParameters())
}
\arguments{
\item{RclusTool.env}{environment in which all global parameters, raw data and results are stored.}
}
\value{
RDS file containing the counts.
}
\description{
Open a Graphical User Interface allowing to choose the images directory, to count manually the number of cells in colonies and to build specific predictive model.
}
\details{
countItemsGUI opens a Graphical User Interface allowing to choose the images directory, to count manually the number of cells in colonies and to build specific predictive model
}
\examples{
\donttest{ 
countItemsGUI()
}
}
\keyword{internal}
