% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleBuilding.R
\name{detailOperation}
\alias{detailOperation}
\title{detail Operation}
\usage{
detailOperation(config = NULL)
}
\arguments{
\item{config}{list of data.frames describing all preprocessing operations.}
}
\value{
newConfig

instr <- rbind(c("select","x","log",""), c("select","y","log",""))
tf2 <- tempfile()
write.table(instr, tf2, sep=",", col.names = FALSE, row.names = FALSE)
operations <- loadPreprocessFile(tf2)

detailOperations<-detailOperation(operations)
}
\description{
detail operation which are in config list
}
\details{
detailOperation develop short operations (combinations select+log for example)
}
\keyword{internal}
