% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RclusTool.R
\name{.logoFrame}
\alias{.logoFrame}
\title{Logo frame in the graphical user interface}
\usage{
.logoFrame(window)
}
\arguments{
\item{window}{window in which the logo frame is positioned.}
}
\value{
None
}
\description{
Create and position the logo frame and images in the graphical user interface.
}
\details{
function to create and position the logo frame in the main window
}
\keyword{internal}
