% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KmeansAutoElbow.R
\name{ElbowFinder}
\alias{ElbowFinder}
\title{Elbow Finder}
\usage{
ElbowFinder(x, y)
}
\arguments{
\item{x}{: x-axis}

\item{y}{: y-axis}
}
\value{
K number of clusters obtained by Elbow method
}
\description{
Perform Elbow method for the automatic estimation of the number of clusters.
}
\details{
ElbowFinder return K number of groups according to Elbow method
}
\examples{

y <- c(824,248,32,28,26,19,20,20,17)           
x <- 1:length(y) 
      
K <- ElbowFinder(x,y)

plot(x,y,type='o')
abline(v=K,col="red")

}
\references{
\url{https://stackoverflow.com/a/42810075}
}
\keyword{internal}
