% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleView.R
\name{ElbowPlot}
\alias{ElbowPlot}
\title{Elbow Plot.}
\usage{
ElbowPlot(nb, method.space.name, RclusTool.env)
}
\arguments{
\item{nb}{tk-notebook in which the Elbow plot will be made.}

\item{method.space.name}{complete name of space.}

\item{RclusTool.env}{: environment in which data and intermediate results are stored.}
}
\value{
None
}
\description{
Display the Elbow Plot after a clustering step.
}
\details{
ElbowPlot displays the Elbow plot after a clustering step.
}
\keyword{internal}
