% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleClustering.R
\name{extractFeaturesFromSummary}
\alias{extractFeaturesFromSummary}
\title{Extraction of features from a summary object.}
\usage{
extractFeaturesFromSummary(summary, split = "...")
}
\arguments{
\item{summary}{a summary object.}

\item{split}{separator string}
}
\value{
list of simple features, features combined to functions, functions combined to features
}
\description{
Extract features from a summary object.
}
\details{
extractFeaturesFromSummary decomposes the summary column names into simple features (not combined to functions), statistic functions and features combined to the functions
}
\examples{

summary <- data.frame("x...min"=1,"x...max"=2,"count"=3,"length"=4)

extractFeaturesFromSummary(summary)

}
\keyword{internal}
