% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleBuilding.R
\name{featSpaceNameConvert}
\alias{featSpaceNameConvert}
\title{Feature Space Name Conversion}
\usage{
featSpaceNameConvert(
  name2convert,
  short2long = TRUE,
  RclusTool.env = initParameters()
)
}
\arguments{
\item{name2convert}{character vector for name to convert.}

\item{short2long}{boolean setting the direction of conversion (higher priority).}

\item{RclusTool.env}{list of parameters.}
}
\value{
character name.
}
\description{
Converts Feature Space Name.
}
\details{
featSpaceNameConvert converts feature space names: either long name to short name, or short name to long name.
}
\examples{

long.name <- "Scaled - Principal Component Analysis"
short.name <- "scaled.pca"

res1 <- featSpaceNameConvert(long.name, short2long=TRUE)
res2 <- featSpaceNameConvert(short.name, short2long=FALSE)

}
\keyword{internal}
