% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleBuilding.R
\name{formatParameterList}
\alias{formatParameterList}
\title{Format Parameter List}
\usage{
formatParameterList(parameterList, lowercase = FALSE)
}
\arguments{
\item{parameterList}{list describing a single preprocessing instruction.}

\item{lowercase}{boolean to build lowercase operation name.}
}
\value{
format Parameter character matrix
}
\description{
formatParameterList format parameters required by following functions, dealing with operations on features
}
\examples{

instr <- list("select","x","log")
formatParameterList <- formatParameterList(instr)

}
\keyword{internal}
