% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleCompute.R
\name{guessFileEncoding}
\alias{guessFileEncoding}
\title{File Encoding Identification.}
\usage{
guessFileEncoding(file.name)
}
\arguments{
\item{file.name}{a character vector, describing one file.}
}
\value{
file character encoding.
}
\description{
Guess file encoding.
}
\details{
guessFileEncoding guess file encoding.
}
\examples{
dat <- rbind(matrix(rnorm(100, mean = 0, sd = 0.3), ncol = 2), 
             matrix(rnorm(100, mean = 2, sd = 0.3), ncol = 2), 
             matrix(rnorm(100, mean = 4, sd = 0.3), ncol = 2))

csv.file <- tempfile()
write.table(dat, csv.file, sep=",", dec=".")

guessFileEncoding(csv.file)


}
\keyword{internal}
