% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchTab.R
\name{initBatchTab}
\alias{initBatchTab}
\title{batch tab}
\usage{
initBatchTab(RclusTool.env, reset = FALSE)
}
\arguments{
\item{RclusTool.env}{: environment in which data and intermediate results are stored.}

\item{reset}{: if TRUE the whole tab is reset, with default options}
}
\value{
None
}
\description{
This function generates the batch tab  of the \code{\link{RclusToolGUI}}, in which the user can batch files.
}
\details{
function to initialize (and to create) the 'batchTab'
}
\keyword{internal}
