% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importTab.R
\name{initImportTab}
\alias{initImportTab}
\title{import tab}
\usage{
initImportTab(RclusTool.env, reset = FALSE)
}
\arguments{
\item{RclusTool.env}{: environment in which data and intermediate results are stored.}

\item{reset}{: if TRUE the whole tab is reset, with default options}
}
\value{
None
}
\description{
This function generates the import tab of the \code{\link{RclusToolGUI}}, in which the user can import files.
}
\details{
function to initialize (and to create) the 'importTab'
}
\keyword{internal}
