% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semisupTab.R
\name{initSemisupTab}
\alias{initSemisupTab}
\title{Semi-Supervised tab}
\usage{
initSemisupTab(RclusTool.env, reset = FALSE)
}
\arguments{
\item{RclusTool.env}{environment in which data and intermediate results are stored.}

\item{reset}{: if TRUE the whole tab is reset, with default options}
}
\value{
None
}
\description{
Generate the semi-supervised classification tab of the \code{\link{RclusToolGUI}}, in which the user can choose and configure the semi-supervised method to apply.
}
\details{
function to initialize (and to create) the 'semisupTab' for data clustering
}
\keyword{internal}
