% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supTab.R
\name{initSupTab}
\alias{initSupTab}
\title{supervised tab}
\usage{
initSupTab(mainWindow, console, graphicFrame, RclusTool.env, reset = F)
}
\arguments{
\item{mainWindow}{: window in which the 'supTab' is created.}

\item{console}{: frame of the RclusTool interface in which messages should be displayed.}

\item{graphicFrame}{: frame of the RclusTool interface in which graphics should be displayed.}

\item{RclusTool.env}{: environment in which data and intermediate results are stored.}

\item{reset}{: if TRUE the whole tab is reset, with default options}
}
\value{
None
}
\description{
This function generates the supervised classification tab of the \code{\link{RclusToolGUI}}, in which the user can choose and configure the classification method to apply.
}
\details{
function to initialize (and to create) the 'supTab' for supervised classification
}
\keyword{internal}
