% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupTab.R
\name{initUnsupTab}
\alias{initUnsupTab}
\title{Unsupervised tab}
\usage{
initUnsupTab(mainWindow, console, graphicFrame, RclusTool.env, reset = F)
}
\arguments{
\item{mainWindow}{: window in which the 'unsupTab' is created.}

\item{console}{: frame of the RclusTool interface in which messages should be displayed.}

\item{graphicFrame}{: frame of the RclusTool interface in which graphics should be displayed.}

\item{RclusTool.env}{: environment in which data and intermediate results are stored.}

\item{reset}{: if TRUE the whole tab is reset, with default options}
}
\value{
None
}
\description{
This function generates the unsupervised classification tab of the \code{\link{RclusToolGUI}}, in which the user can choose and configure the clustering method to apply.
}
\details{
function to initialize (and to create) the 'unsupTab' for data clustering
}
\keyword{internal}
