% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleBuilding.R
\name{listDerivableFeatureSpaces}
\alias{listDerivableFeatureSpaces}
\title{Builds list of derivable feature spaces}
\usage{
listDerivableFeatureSpaces(
  scaling = FALSE,
  pca = FALSE,
  spectral = FALSE,
  RclusTool.env = initParameters()
)
}
\arguments{
\item{scaling}{boolean TRUE if scaling required.}

\item{pca}{boolean TRUE if Principal Components Analysis required.}

\item{spectral}{boolean TRUE if Spectral Embedding required.}

\item{RclusTool.env}{environment in which all global parameters, raw data and results are stored.}
}
\value{
character vector naming feature spaces.
}
\description{
Builds list of derivable feature spaces.
}
\details{
listDerivableFeatureSpaces build the list of feature spaces matching parameters.
}
\examples{

listDerivableFeatureSpaces(scaling=TRUE, spectral=TRUE)

}
\keyword{internal}
