% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RclusTool.R
\name{MainWindow}
\alias{MainWindow}
\title{Main window}
\usage{
MainWindow(RclusTool.env)
}
\arguments{
\item{RclusTool.env}{environment in which data and results will be stored. If NULL, a local environment will be created.}
}
\value{
None
}
\description{
Generate an user-friendly interface to classify data in a unsupervised, semi-supervised or supervised way.
}
\details{
function to display the graphical user interface to classify dataset.
}
\keyword{internal}
