% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleView.R
\name{makeTitle}
\alias{makeTitle}
\title{RclusTool makeTitle.}
\usage{
makeTitle(string, length = 120, prefix.length = 5, char = "_")
}
\arguments{
\item{string}{string to be completed}

\item{length}{final title length in characters}

\item{prefix.length}{space length in front of the title}

\item{char}{character used to complete the title}
}
\value{
title
}
\description{
Makes a character title from string by completion with a specified character.
}
\details{
makeTitle Makes a title from a string.
}
\keyword{internal}
