% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleBuilding.R
\name{matchNames}
\alias{matchNames}
\title{Match Names}
\usage{
matchNames(requested.names, names.set)
}
\arguments{
\item{requested.names}{a character vector from an operation.}

\item{names.set}{a character vector from Features/Profiles column.}
}
\value{
character
}
\description{
make a new character from two characters with maj/min adjustment
}
\details{
matchNames correct Maj/Min to match operations and Features/Profiles columns
}
\examples{

matchedName<-matchNames("Profiles","profiles")

matchedName

}
\keyword{internal}
