% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleView.R
\name{messageConsole}
\alias{messageConsole}
\title{RclusTool consoleMessage.}
\usage{
messageConsole(message, RclusTool.env = initParameters())
}
\arguments{
\item{message}{string to print in the console.}

\item{RclusTool.env}{environment in which all global parameters, raw data and results are stored.}
}
\value{
None
}
\description{
Adds a message in the RclusTool GUI.
}
\details{
consoleMessage Adds a message in the RclusTool GUI.
}
\keyword{internal}
