% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importTab.R
\name{previewCSVfile}
\alias{previewCSVfile}
\title{Preview CSV file}
\usage{
previewCSVfile(file, sep, dec, na.strings, RclusTool.env, nrows = 3, ...)
}
\arguments{
\item{file}{: csv filename.}

\item{RclusTool.env}{environment in which data and intermediate results are stored.}

\item{nrows}{: number of rows to be read.}

\item{...}{: other parameters of read.csv function.}
}
\value{
None
}
\description{
This function generates a window printing first rows of the csv file and the first rows of the data.frame obtained.
}
\details{
function to preview csv file
}
\keyword{internal}
