% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RclusTool.R
\name{RclusToolGUI}
\alias{RclusToolGUI}
\title{Username and user type selection}
\usage{
RclusToolGUI(RclusTool.env = new.env(), debug = FALSE)
}
\arguments{
\item{RclusTool.env}{environment in which data and results will be stored. If NULL, a local environment will be created.}

\item{debug}{boolean: if TRUE, the debug mode is activated.}
}
\value{
Nothing, just open the graphical user interface.
}
\description{
Generate a first window to enter the username and to select the user type ('standard' or 'expert').
}
\details{
function to display the first window of the RclusTool interface (username and user type selection)
}
\examples{
RclusToolGUI()

}
