% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleClustering.R
\name{saveLogFile}
\alias{saveLogFile}
\title{Log file saving}
\usage{
saveLogFile(filename.txt, txt, dir)
}
\arguments{
\item{filename.txt}{character vector specifying the path and the name of the txt file.}

\item{txt}{character vector describing the different steps of the analyses.}

\item{dir}{character vector specifying the directory where to save the txt file.}
}
\description{
Save a log txt file at the end of the session, describing the different steps of the analyses.
}
\details{
saveLogFile saves a log txt file at the end of the session, describing the different steps of the analyses
}
\examples{
logfile <- 'logfile.txt'
saveLogFile(logfile, txt=rbind("Analysis date: ...", "Analysis duration: ..."), tempdir())


}
\keyword{internal}
