% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleBuilding.R
\name{savePreprocess}
\alias{savePreprocess}
\title{Preprocessing exportation}
\usage{
savePreprocess(filename.csv, config, dir)
}
\arguments{
\item{filename.csv}{character vector specifying the name of the csv file.}

\item{config}{4-columns character matrix describing all preprocessing operations.}

\item{dir}{character vector specifying the directory of the csv file.}
}
\value{
csv file containing preprocessing.
}
\description{
Export all preprocessing operations in a csv file.
}
\details{
savePreprocess exports all preprocessing operations in a csv file
}
\examples{
test.file <- tempfile()
config <- matrix(c("select","x",NA,NA,"select","y",NA,NA), byrow=TRUE, ncol=4)
savePreprocess(basename(test.file), config, dirname(test.file))


}
