% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeSpectral.R
\name{search.neighbour}
\alias{search.neighbour}
\title{Search neighbour}
\usage{
search.neighbour(vdist, k)
}
\arguments{
\item{vdist}{distance vector of point with other points}

\item{k}{number of neighbours}
}
\value{
id of the Kth nearest neighbour
}
\description{
search.neighbour give the id of the Kth nearest neighbour
}
\examples{

vdist=c(0,2,1,5,7,1,3)

id<-search.neighbour(vdist,2)
}
\seealso{
\code{\link{computeGaussianSimilarityZP}}
}
\keyword{internal}
