% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleCompute.R
\name{sortCharAsNum}
\alias{sortCharAsNum}
\title{Character vector numeric sorting}
\usage{
sortCharAsNum(char.vec, ...)
}
\arguments{
\item{char.vec}{a character vector, describing numbers.}

\item{...}{other parameters controlling the order treatment of the character vector.}
}
\value{
the character vector sorted, if possible.
}
\description{
Performs sorting of a vector of characters numerically, if possible
}
\details{
sortCharAsNum performs sorting of a vector of characters numerically, if possible
}
\examples{
x <- c("12", "8", "101")

sortCharAsNum(x)

}
\seealso{
\code{\link{computeSemiSupervised}}, \code{\link{KwaySSSC}}
}
\keyword{internal}
