% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleView.R
\name{StringToTitle}
\alias{StringToTitle}
\title{String To Title.}
\usage{
StringToTitle(string, size, fontsize = 11)
}
\arguments{
\item{string}{: string to convert}

\item{size}{: size of the title in cm}

\item{fontsize}{: size of font (default=11)}
}
\value{
title
}
\description{
Make a grahical title from string.
}
\details{
StringToTitle Convert a string to title.
}
\keyword{internal}
