% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleView.R
\name{tk2add.notetab}
\alias{tk2add.notetab}
\title{Add notetab.}
\usage{
tk2add.notetab(nb, tab.label, tab.name = NULL)
}
\arguments{
\item{nb}{a notebook.}

\item{tab.label}{string the label of a tab on the interface.}

\item{tab.name}{string the name of the tab in a notebook.}
}
\value{
None
}
\description{
Add a notetab.
}
\details{
tk2add.notetab add a notetab.
}
\keyword{internal}
