% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleView.R
\name{tk2delete.notetab}
\alias{tk2delete.notetab}
\title{Delete notetab inside a tk-notebook}
\usage{
tk2delete.notetab(nb, tab.names = NULL)
}
\arguments{
\item{nb}{a notebook.}

\item{tab.names}{either NULL to delete all the notetabs, or a vector of notetab names.}
}
\value{
None
}
\description{
Delete notetab inside a tk-notebook.
}
\details{
tk2delete.notetab delete a notetab
}
\keyword{internal}
