% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleView.R
\name{tk2notetab.RclusTool}
\alias{tk2notetab.RclusTool}
\title{RclusTool tk2notetab.}
\usage{
tk2notetab.RclusTool(nb, tab.label, tab.name = NULL)
}
\arguments{
\item{nb}{a tk2notebook or ttk2notebook widget.}

\item{tab.label}{the label of a tab on the interface.}

\item{tab.name}{the name of the tab in a notebook.}
}
\value{
None
}
\description{
RclusTool adaptation of tk2notetab; builds and returns a R tk2notetab matching a notetab already declared in tk notebook.
}
\details{
tk2notetab.RclusTool RclusTool adaptation of tk2notetab.
}
\keyword{internal}
