% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleView.R
\name{tkEmptyLine}
\alias{tkEmptyLine}
\title{RclusTool tkEmptyLine.}
\usage{
tkEmptyLine(tk.object, row = NULL)
}
\arguments{
\item{tk.object}{tk object in which to insert the new line}

\item{row}{row number of the insertion}
}
\value{
None
}
\description{
Makes a character title from string by completion with a specified character.
}
\details{
tkEmptyLine Inserts an empty line in a graphical tk objects dealt with grid.
}
\keyword{internal}
