% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleView.R
\name{tkrplot.RclusTool}
\alias{tkrplot.RclusTool}
\title{RclusTool tkrplot.}
\usage{
tkrplot.RclusTool(graphicFrame, fun, hscale = 1, vscale = 1)
}
\arguments{
\item{graphicFrame}{frame of the RclusTool interface in which graphics should be displayed.}

\item{fun}{a function}

\item{hscale}{numeric value corresponding to the horizontal scale of graphic}

\item{vscale}{vertical scale}
}
\value{
None
}
\description{
RclusTool adaptation of tkrplot.
}
\details{
tkrplot.RclusTool RclusTool adaptation of tkrplot.
}
\keyword{internal}
