% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleView.R
\name{tkrreplot.RclusTool}
\alias{tkrreplot.RclusTool}
\title{RclusTool tkrreplot.}
\usage{
tkrreplot.RclusTool(env.graphic)
}
\arguments{
\item{env.graphic}{graphic environment generated by tkrplot.RclusTool}
}
\value{
None
}
\description{
RclusTool adaptation of tkrreplot.
}
\details{
tkrreplot.RclusTool RclusTool adaptation of tkrreplot.
}
\seealso{
\code{\link{tkrreplot.RclusTool}}
}
\keyword{internal}
