% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleCompute.R
\name{toStringDataFrame}
\alias{toStringDataFrame}
\title{To String Data Frame}
\usage{
toStringDataFrame(object, digits = NULL)
}
\arguments{
\item{object}{dataframe to convert to string}

\item{digits}{digits in dataframe}
}
\value{
string to print.
}
\description{
Convert dataframe to string to print it in console.
}
\details{
toStringDataFrame convert dataframe to string to print it in console
}
\references{
\url{https://stackoverflow.com/a/45541857}
}
\keyword{internal}
