% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importTab.R
\name{buildImportTab}
\alias{buildImportTab}
\title{Build Import tab}
\usage{
buildImportTab(RclusTool.env)
}
\arguments{
\item{RclusTool.env}{environment in which data and intermediate results are stored.}
}
\value{
None
}
\description{
Generate the files importation tab of the \code{\link{RclusToolGUI}}, in which the user can select original files (features, metadata, signals or images).
}
\details{
function to create the 'importTab' for data importation (features, metadata, signals and images)
}
\keyword{internal}
