% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semisupTab.R
\name{buildSemisupTab}
\alias{buildSemisupTab}
\title{Semi-Supervised tab}
\usage{
buildSemisupTab(mainWindow, console, graphicFrame, RclusTool.env)
}
\arguments{
\item{mainWindow}{window in which the 'semisupTab' is created.}

\item{console}{frame of the RclusTool interface in which messages should be displayed.}

\item{graphicFrame}{frame of the RclusTool interface in which graphics should be dispayed.}

\item{RclusTool.env}{environment in which data and intermediate results are stored.}
}
\description{
Generate the semi-supervised classification tab of the \code{\link{RclusToolGUI}}, in which the user can choose and configure the semi-supervised method to apply.
}
\details{
function to build the 'semisupTab' for semi-supervised classification
}
\keyword{internal}
