% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupTab.R
\name{buildUnsupTab}
\alias{buildUnsupTab}
\title{Unsupervised tab}
\usage{
buildUnsupTab(mainWindow, console, graphicFrame, RclusTool.env)
}
\arguments{
\item{mainWindow}{: window in which the 'unsupTab' is created.}

\item{console}{: frame of the RclusTool interface in which messages should be displayed.}

\item{graphicFrame}{: frame of the RclusTool interface in which graphics should be displayed.}

\item{RclusTool.env}{: environment in which data and intermediate results are stored.}
}
\description{
This function generates the unsupervised classification tab of the \code{\link{RclusToolGUI}}, in which the user can choose and configure the clustering method to apply.
}
\details{
function to create the 'unsupTab' for data clustering
}
\keyword{internal}
