% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleView.R
\name{ElbowPlot}
\alias{ElbowPlot}
\title{Elbow Plot.}
\usage{
ElbowPlot(nb, method.space.name, RclusTool.env, hscale = 1.2, charsize = 11)
}
\arguments{
\item{nb}{tk-notebook in which the Elbow plot will be made.}

\item{method.space.name}{complete name of space.}

\item{RclusTool.env}{environment in which data and intermediate results are stored.}

\item{hscale}{numeric value corresponding to the horizontal scale of graphic}

\item{charsize}{character size}
}
\value{
None
}
\description{
Display the Elbow Plot after a clustering step.
}
\details{
ElbowPlot displays the Elbow plot after a clustering step.
}
\keyword{internal}
