% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importTab.R
\name{initImportTab}
\alias{initImportTab}
\title{import tab}
\usage{
initImportTab(mainWindow, console, graphicFrame, RclusTool.env, reset = F)
}
\arguments{
\item{mainWindow}{: window in which the 'importTab' is created.}

\item{console}{: frame of the RclusTool interface in which messages should be displayed.}

\item{graphicFrame}{: frame of the RclusTool interface in which graphics should be displayed.}

\item{RclusTool.env}{: environment in which data and intermediate results are stored.}

\item{reset}{: if TRUE the whole tab is reset, with default options}
}
\value{
None
}
\description{
This function generates the import tab of the \code{\link{RclusToolGUI}}, in which the user can import files.
}
\details{
function to initialize (and to create) the 'importTab'
}
\keyword{internal}
