% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessTab.R
\name{initPreprocessTab}
\alias{initPreprocessTab}
\title{build Preprocess tab}
\usage{
initPreprocessTab(mainWindow, console, graphicFrame, RclusTool.env,
  reset = F, readConfig = F)
}
\arguments{
\item{mainWindow}{window in which the 'preprocessTab' is created.}

\item{console}{frame of the RclusTool interface in which messages should be displayed.}

\item{graphicFrame}{frame of the RclusTool interface in which graphics should be dispayed.}

\item{RclusTool.env}{environment in which data and intermediate results are stored.}

\item{reset}{boolean to reset the whole interface}

\item{readConfig}{boolean to reset the whole interface according the config parameters of the data.sample}
}
\description{
Generate the data preprocessing tab of the \code{\link{RclusToolGUI}}, in which the user can select, transform, filter or create variables.
}
\details{
function to initialize/refresh the 'preprocessTab' for data preprocessing (variables selection, transformation, creation, ...)
}
\keyword{internal}
