% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semisupTab.R
\name{initSemisupTab}
\alias{initSemisupTab}
\title{Semi-Supervised tab}
\usage{
initSemisupTab(mainWindow, console, graphicFrame, RclusTool.env,
  reset = F)
}
\arguments{
\item{mainWindow}{window in which the 'semisupTab' is created.}

\item{console}{frame of the RclusTool interface in which messages should be displayed.}

\item{graphicFrame}{frame of the RclusTool interface in which graphics should be dispayed.}

\item{RclusTool.env}{environment in which data and intermediate results are stored.}

\item{reset}{: if TRUE the whole tab is reset, with default options}
}
\description{
Generate the semi-supervised classification tab of the \code{\link{RclusToolGUI}}, in which the user can choose and configure the semi-supervised method to apply.
}
\details{
function to initialize (and to create) the 'semisupTab' for data clustering
}
\keyword{internal}
