% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleView.R
\name{tk2draw.notetab}
\alias{tk2draw.notetab}
\title{Draw in a Notetab.}
\usage{
tk2draw.notetab(nb, tab.name, fun, hscale = 1.2)
}
\arguments{
\item{nb}{a notebook.}

\item{tab.name}{name of the tk2notetab variable.}

\item{fun}{a function.}

\item{hscale}{numeric value corresponding to the horizontal scale of graphic}
}
\value{
None
}
\description{
Draw in a Notetab with a function.
}
\details{
tk2draw.notetab draw in a notetab.
}
\keyword{internal}
